Adds default template arguments to expansion
When a fixture is created the location of the artifacts is dynamic.
This means that file paths and pointers within artifacts can't be
set in the fixture until it runs.  A new default DIR parameter was added
that can be used within a template to get this directory when the
fixture is generated.