Coverts from constructor function to class
This simplifies the cleanup procedure.  It only needs to be called once
now at the end of tests.  The old method meant that one needed to call
it after each test.  If it were forgotten, then the artifacts could be
left in the temp location.  Now the class keeps track of all base
directories used in the tests and removes them all at the end with one
static function.

- Changed cleanup() to a static function in the class
- Updated documentation to reflect change in cleanup procedure