Fixes relative import problem with package.json
When a module uses the Fixture it would expect that the overrides in its
package.json to be applied.  This class was using a relative import.
This meant that instead of using the modules overrides it was always
using the Fixture modules override.  Changed this to use the cwd() of
the executing module to find package.json.